+++ Viewer/src/VGenVarAttr.cpp
@@ -87,8 +87,8 @@ void VGenVarAttr::scan(VNode* vnode,std::vector<VAttribute*>& vec)
     {
         std::vector<Variable> v;
         vnode->genVariables(v);
-        int n=v.size();
-        for(size_t i=0; i < n; i++)
+        int n=static_cast<int>(v.size());
+        for(int i=0; i < n; i++)
         {
             vec.push_back(new VGenVarAttr(vnode,v[i],i));
         }
